Fixed some bugs and moved TODO.md to issues.

+++ include/processor.h
@@ -3,6 +3,6 @@
 #ifndef _PROCESSOR_H_
 #define _PROCESSOR_H_
 bool isDelimiter(char c);
-string getNextToken(int index, string buffer);
+string getNextToken(size_t index, string buffer);
 void process(string buffer);
 #endif
